Добавление наследников ViewModel и Fragment для редактирования элемента Playlist
