+++ app/src/main/java/com/example/playlistmaker/presentation/playlists/edit/PlaylistEditViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.presentation.playlists.edit
+
+import com.example.playlistmaker.domain.playlists.api.PlaylistEditInteractor
+
+class PlaylistEditViewModel(playlistEditInteractor: PlaylistEditInteractor) :
+    PlaylistCreateViewModel(playlistEditInteractor) {
+}
