+++ app/src/main/res/layout/fragment_playlist_edit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.playlists.edit.PlaylistEditFragment">
+    tools:context=".ui.playlists.edit.PlaylistCreateFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/topPanelBottomLine"
